refactor: space layout directory and naming
