+++ app/spaces/[spaceid]/components/sidebar/spaceSidebar.tsx
@@ -68,7 +68,7 @@ const Item = ({ item }: { item: SidebarItem }) => {
   );
 };
 
-const MainSidebar = () => {
+const SpaceSidebar = () => {
   return (
     <div className="mobile:hidden fixed top-[50px] bottom-0 left-0 w-[62px] h-[calc(100vh-50px)] bg-[#2C2C2C] pt-2.5 pb-2.5 g-2.5 border-r border-[rgba(255,255,255,0.1)] ">
       <div className="w-full flex flex-col items-center pb-2.5 gap-2.5 border-b border-[rgba(255,255,255,0.1)]">
@@ -81,4 +81,4 @@ const MainSidebar = () => {
   );
 };
 
-export default MainSidebar;
+export default SpaceSidebar;
