+++ app/spaces/[spaceid]/components/spaceLayout.tsx
@@ -1,9 +1,9 @@
 'use client';
 
 import { useSelectedLayoutSegments } from 'next/navigation';
-import MainSubSidebar from '@/app/spaces/[spaceid]/components/sidebar/subSidebar/mainSubSidebar';
 import React from 'react';
-import MainSidebar from '@/app/spaces/[spaceid]/components/sidebar/mainSidebar';
+import SpaceSidebar from '@/app/spaces/[spaceid]/components/sidebar/spaceSidebar';
+import SpaceSubSidebar from '@/app/spaces/[spaceid]/components/sidebar/spaceSubSidebar/spaceSubSidebar';
 
 const SpaceLayout = () => {
   const segments = useSelectedLayoutSegments();
@@ -11,11 +11,11 @@ const SpaceLayout = () => {
 
   return (
     <>
-      <MainSidebar />
+      <SpaceSidebar />
 
       {isSettingPage ? null : (
         <div className="mobile:hidden tablet:hidden">
-          <MainSubSidebar />
+          <SpaceSubSidebar />
         </div>
       )}
     </>
